ai-engine/requirements.txt
# Zara ETL Production Requirements
# =================================
# Auto-generated from working venv with latest versions
# Python 3.13 compatible
# Generated: $(date +%Y-%m-%d)
#
# To regenerate:
# 1. Create venv: python3 -m venv temp_venv
# 2. Activate: source temp_venv/bin/activate  
# 3. Install core: pip install apache-airflow apache-airflow-providers-{postgres,http,ftp}
# 4. Install rest: pip install [all other packages]
# 5. Freeze: pip freeze > requirements.txt
# =================================

a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==4.0.2
aiosqlite==0.21.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
apache-airflow==3.1.0
apache-airflow-core==3.1.0
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-ftp==3.13.2
apache-airflow-providers-http==5.3.4
apache-airflow-providers-postgres==6.3.0
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.0
apache-airflow-task-sdk==1.1.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arxiv==2.2.0
asgiref==3.9.2
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
backrefs==5.9
bcrypt==5.0.0
billiard==4.2.2
boto3==1.40.41
botocore==1.40.41
cadwyn==5.4.4
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
feedparser==6.0.12
flower==2.0.1
frozenlist==1.7.0
fsspec==2025.9.0
ghp-import==2.1.0
googleapis-common-protos==1.70.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanize==4.13.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lazy-object-proxy==1.12.0
libcst==1.8.5
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mergedeep==1.3.4
methodtools==0.4.7
minio==7.2.18
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.20
mkdocs-material-extensions==1.3.1
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.6.4
natsort==8.4.0
numpy==2.3.3
openai==1.109.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
outcome==1.3.0.post0
packaging==25.0
paginate==0.5.7
pandas==2.3.3
passlib==1.7.4
pathlib_abc==0.5.1
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pymdown-extensions==10.16.1
PyMuPDF==1.26.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
PyYAML-ft==8.0.0
pyyaml_env_tag==1.1
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
retryhttp==1.3.3
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
sentry-sdk==2.39.0
setproctitle==1.3.7
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.48.0
structlog==25.4.0
supervisor==4.3.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tiktoken==0.11.0
tornado==6.5.2
tqdm==4.67.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.3.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.37.0
uvloop==0.21.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
wirerope==1.0.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
xmltodict==1.0.2